+++ src/obfs/tls1.2_ticket.h
@@ -14,7 +14,7 @@ void tls12_ticket_auth_dispose(struct obfs_t *obfs);
 size_t tls12_ticket_auth_client_encode(struct obfs_t *obfs, char **pencryptdata, size_t datalength, size_t* capacity);
 ssize_t tls12_ticket_auth_client_decode(struct obfs_t *obfs, char **pencryptdata, size_t datalength, size_t* capacity, int *needsendback);
 
-int tls12_ticket_auth_get_overhead(struct obfs_t *obfs);
+size_t tls12_ticket_auth_get_overhead(struct obfs_t *obfs);
 
 struct buffer_t * tls12_ticket_auth_server_pre_encrypt(struct obfs_t *obfs, struct buffer_t *buf);
 struct buffer_t * tls12_ticket_auth_server_encode(struct obfs_t *obfs, struct buffer_t *buf);
@@ -27,7 +27,7 @@ bool tls12_ticket_auth_server_udp_post_decrypt(struct obfs_t *obfs, struct buffe
 
 void * tls12_ticket_fastauth_init_data(void);
 void tls12_ticket_fastauth_new_obfs(struct obfs_t *obfs);
-int tls12_ticket_fastauth_get_overhead(struct obfs_t *obfs);
+size_t tls12_ticket_fastauth_get_overhead(struct obfs_t *obfs);
 void tls12_ticket_fastauth_dispose(struct obfs_t *obfs);
 size_t tls12_ticket_fastauth_client_encode(struct obfs_t *obfs, char **pencryptdata, size_t datalength, size_t* capacity);
 ssize_t tls12_ticket_fastauth_client_decode(struct obfs_t *obfs, char **pencryptdata, size_t datalength, size_t* capacity, int *needsendback);
